Fix for incorrect error message in tm03.
Also adds a new enum value to trGridType: triangularmesh; this will
only be useful when the triangular mesh code is checked in.

Changes the default value of the trOutOfRangeF resource from 1e12 to
1e30. This allows the lower level routines to distinguish between a
value that is outside the data grid and one that falls outside the limb
of a map projection.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@7284 10fe2890-8a40-4696-8b1d-7ccbb19c398e