Added a new function "gsn_legend_ndc_return". This is identical to "gsn_legend_ndc", except it returns a value. In fact, gsn_legend_ndc just calls this function, but doesn't return anything.
This function partially addresses an issue that a user was having with
adding a legend to a plot.  I may eventually add another procedure called
"add_annotation" that allows a user to add an annotation to a plot in
which position (0,0) represents the dead center of a plot, and then the
user can move the annotation up or down and/or left or right. This is
a nice way to add annotations to a plot, because they get attached to
the plot, and you don't have to fiddle with NDC coordinates or plot
coordinates to get the annotation where you want it.  An annotation
can be any other drawable object, like a text string, a labelbar, a
legend, even another plot.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@7308 10fe2890-8a40-4696-8b1d-7ccbb19c398e