Added zlib to the list of subdirectories.
Please note: the zlib source is going to be added in such a way that
it will get built in with the GKS library. This is easier for
development and testing purpsoes, and it avoids making the user having
to add another library to their compile line for their NCAR Graphics
programs.

Eventually, we may want to rethink this. This library could somehow
conflict with the zlib that is included with HDF. However, the latest
version of HDF (which I have not downloaded yet) no longer includes
this library...


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@7356 10fe2890-8a40-4696-8b1d-7ccbb19c398e