This is a first cut at the PIX workstation driver. There are still some loose ends, but it works now to output the X window dump format, XWD. I created a dummy PIX_Write_PNG function to parallel the implemented PIX_Write_XWD function.
The filename is supplied as part of the NGC_PIXCONFIG C escape function.
The filename is treated as a basename, where the first frame is given
the name <filename>.xwd, and subsequent frames are given names:
<filename>.1.xwd
<filename>.2.xwd
...
<filename>.n.xwd

I need to clean up the error handling and some of the structures could
be simplified.
By changing just 2 places in the pixcontrol.c file, you can enable an
X window that draws a copy of the PIXMAP that is created. I have left
this a a possibly useful debugging aid.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@7363 10fe2890-8a40-4696-8b1d-7ccbb19c398e