This is a workaround fix for Grib files from the IPCC Data Distribution Center. (http://cera-www.dkrz.de/IPCC_DDC/)
Several unrecognized centers are used for these files. The work-around code
is activated when an unrecognized center is encountered. I have only looked at
a limited number of these files, so I do not know how comprehesive this fix
is.

User-defined parameter tables are still needed and I do not attempt to supply
a name for the center.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@7386 10fe2890-8a40-4696-8b1d-7ccbb19c398e