This is a fix for remedy TT #046175 that Sylvia reported. If a user set gsnMaximize to True for a PS or PDF workstation (which causes the device coordinates to get calculated such that the plot is maximized on the specified paper size), and then a second plot *without* setting gsnMaximize, then the old device coordinates would still be in effect, and might cause the plot to look drastically wrong. This is especially true if the user first draws a paneled plot with gsnMaximize set to True, and then draws a single plot with gsnMaximize set to False.
The device coordinates will only be set back if the user lets
the plotting routine advance the frame, because the coordinates can
only be set back to their original values *after* the frame
is advanced. (Otherwise, the original values will overwrite the
calculated values.)

If the user turns off gsnFrame, then he/she will be responsible for
making sure the next plot drawn to the PS/PDF workstation is done so
with correct device coordinates. (This can be accomplished by setting
gsnMaximize to True, or else setting the device coordinates yourself.)

This change will need lots of testing!


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@7388 10fe2890-8a40-4696-8b1d-7ccbb19c398e