Added a new procedure "reset_device_coordinates" that will reset the PS/PDF device coordinates and the orientation back to the default. This function is now called in the "draw_and_frame" procedure, but only if the frame is advance.
Also, when gsn_open_wks is called with "ps" or "pdf" output, the
device coordinates and the orientation values are retrieved, and
attached to the return variable "wks" as attributes.

Then, when reset_device_coordinates is called with "wks" as an
argument, it checks to see that these attributes are set. If so, then
they are used to reset the device coordinates and the orientation.
If not, then these values are set to the NCL defaults.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@7391 10fe2890-8a40-4696-8b1d-7ccbb19c398e