Added "reset_device_coordinates" to the "maximize_output" function (this function is internal.)
SIGNIFICANT CHANGE: for some reason, the longitude labels and the
gsnxxxString fonts were helvetica-bold for the polar plots, which is
inconsistent with the other plot templates that all use
helvetica. Sylvia, Adam, and I discussed this, and decided that these
labels should be helvetica like everybody else.

The main title font will still be helvetica-bold.

If one wants to keep helvetica-bold for the polar label fonts, they
can set gsnPolarLabelFont to "helvetica-bold".

Same for gsnRight/Left/CenterString. They can set gsnStringFont to
"helvetica-bold" if they want to keep the font bold.

THIS NEEDS TO BE ADVERTISED!!  I will put it in the "what's new in
NCL" list.  It should also go in the polar plot section, and maybe the
titles section.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@7397 10fe2890-8a40-4696-8b1d-7ccbb19c398e