Addresses TT#047704 and TT#045877:
Two additional adjunct time variables are now created when GRIB files are
read for each instance of initial_time(n). This is an attempt to mitigate
the fact NCL is not COARDS-compliant with respect to its initial_time coordinate
variable, because it is a string and therefore cannot be considered numerically
monotonic. The first variable added would be a legitimate time coordinate
variable, if it could replace the string coordinate variable. Its name is formed
from the corresponding initial_time(n) variable as follows, where (n) represents
the integer dimension number:
   initial_time(n)_hours
and always has units of "hours since 1800-01-01 00:00"
The second variable is for convenience only and is the date encoded as a double.
Its name is of the form:
   initial_time(n)_encoded
and its units are "yyyymmddhh.hh_frac"

Suffixes for supported file formats are now processed case-insensitively such that "nc", "NC",
"nC", and "Nc" are now all legitimate suffixes for netCDF files. This change applies to
all recognized suffixes.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@7492 10fe2890-8a40-4696-8b1d-7ccbb19c398e