Added get_ncl_version (returns the NCL version).
Also, so I would lose these functions by accident, I added
ctwrap.f and ctwrapW.c. This is just a wrapper that calls one of
Dave K's examples for generating contours on a SEAM grid.
Eventually we'll remove this function when the SEAM grid becomes
recognized by NCL directly.  This is not to be an advertised function.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@7502 10fe2890-8a40-4696-8b1d-7ccbb19c398e