Preliminary check in of Mesh grid code for people to test. The main interface change is that now there is a new ScalarField subclass called MeshScalarField. It can be called in three ways: (1) using an ElementNode array that specifies the indexes into an array of points to define a set of triangles -- this is like the chesapeake bay example. (2) using a auxiliary cell boundary areas, from which the triangles are deduced by sorting the cell vertices. (3) with only a 1d set of x and y coordinates, in which case Shewchuk's Triangle code is applied to determine a Delaunay triangulation of the points.
MeshScalarField uses sfDataArray, sfXArray, and sfYArray, but unlike
the regular ScalarField these must be one-dimensional.
It also defines these new resources:
sfElementNodes: which is a 2D integer array dimensions num_triangles x 3
whose values are indexes into sfXArray and sfYArray.
sfFirstNodeIndex which basically is used to say whether you are numbering
the indexes from 1 or from 0.
sfXCellBounds and sfYCellBounds which is dimensioned number_of_cells by
(usually) 6, and is the way the geodesic examples are done.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@7576 10fe2890-8a40-4696-8b1d-7ccbb19c398e