Okay, everybody needs to pay attention to this CVS check-in!
I've had to create another macro called NGCALLC to handle the
referencing of common blocks variables from C routines.

Before, we were just using the NGCALLF macro, but this was based on the
assumption that Fortran routine names and common block names would be
handled the same way by the compiler.

This is not the case with the Absoft compiler, which more people are
starting to use (as you saw in ncarg-talk). I created a separate
macro to use only with common blocks.

I believe there are only five places in the NCARG code where
C routines are referencing Fortran common blocks routines:

     Three C examples: c_cbex01.c, c_srex01.c, c_vvex03.c
     Two awiC routines: s_gopwk.c and s_gesc.c

I will be checking in the mods to these five routines shortly.

I just realized that there may be places in the HLU/NCL code that
will need attention, so I'll look into that next.

--Mary


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@7611 10fe2890-8a40-4696-8b1d-7ccbb19c398e