Now that the options for compiling NCAr Graphics using AbsoftProFortran have been figured out, I decided to get rid of the confusing Darwin.f77 and Darwin.f90 and just make a generic Darwin.absoft.
Also, I no longer need all those weird options for forcing the Fortran
routine names to be lowercase, and for the common block names to be
lowercase *and* not have a "_C" appended.  I work around these via
the new NGCAPS and NGCALLC macros.

Also, and this is important, you want to use gcc as your compiler and
not the generic "cc".


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@7615 10fe2890-8a40-4696-8b1d-7ccbb19c398e