On some systems, there's no case-sensitivity in file names, so file.f and file.F would be considered the same.
This was causing problems for the special cases where we had *.F files that
have to be run through sed first. The *.F file was effectively getting
overwritten to a zero-length filee.

I'm renaming the *.F files to *.f.sed, and changing the FCPP rule in Template
to reflect this.

--Mary


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@7617 10fe2890-8a40-4696-8b1d-7ccbb19c398e