I have modified this example to use the CONPACKT routine CTTMTX instead of CTTMTL to construct the triangular meshes.  I have done this because, on one of the machines Mary is testing the example on, it failed in such a way as to make me think that this is necessary.  We'll see ...
Mary, this is with reference to the system used for NCGMs in your directory
"macosx.gnu"; "ctex02" failed with the following error message:

  ERROR    1 IN ICAEDG - LOGIC ERROR

See what happens when you try this new version and let me know.  You should
also retry this example on some other machines to make sure it still works
on them.  I've tested it on my Sun workstation, but that's it ...


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@7645 10fe2890-8a40-4696-8b1d-7ccbb19c398e