Added a temporary work-around in the meshScalarFieldClass that sets the sf*CellBounds resources in the "create" call rather than in the "setvalues" call.
Removed restriction from gsn_csm contour scripts that "data" be a 2D array.

With this check-in, the new triangulation stuff done by Dave K and B
is available from the gsn scripts. I have tested the gsn_csm_contour
and gsn_csm_contour*map routines, but not the ones that involve overlaying
vectors and streamlines.

Also, I noticed that in all of Dave B's examples for mesh scalar fields
which he overlays on a map, he retrieves the same four sf*Actual*F resources.

I may make this an automatic part of the gsn_csm_contour_map* routines
so the user doesn't have to do it.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@7678 10fe2890-8a40-4696-8b1d-7ccbb19c398e