I have made a small change in the math of the routine CPTROE to fix a problem uncovered on the Linux system "longs".  This should enable us to use the same level of optimization for that routine as for the rest.
I'm still not sure why turning on the optimizer caused this problem to
surface.  I'm guessing that the level of optimization used allowed the
compiler to substitute mathematically equivalent expressions for ones
implied by the code and this somehow caused a quantity that was exactly
zero to become slightly non-zero.  In any case, the code, as written,
was not guaranteed to handle certain cases and needed to be fixed, so
it's a good thing the problem was uncovered, however it happened.

The same fix will be put in the routine CTTROE, in CONPACKT.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@7760 10fe2890-8a40-4696-8b1d-7ccbb19c398e