This fixes some problems with NCL HDF File I/O. 1) Variable names now have non-alphanumeric characters replaced with an underscore. Then multiple underscores are collapsed to a single underscore. This brings the HDF code in line with what has always been documented.
Each variable now also is given an attribute called "hdf_name" whose value is the
original HDF variable name. This may often be a redundant piece of information, but
it seems better to always have it so that the user code does not have to check whether
it exists.

2) Certain HDF files have SDS variables with identical names. They are distinguished
in the HDF file by belonging to different hierarchically arranged 'Vgroups', rather
like a directory structure in a file system. These are now distinguished by apppending
a double underscore followed by the Vgroup reference id to the variable name.
In addition a text representation of the Vgroup hierarchy (with '/' used as a
separator is added as the attribute 'hdf_group'. Also the group reference id
is stored as an integer attribute called 'hdf_group_id'. Here's an example fragment
of calling print on a file variable:

      float Longwave_flux_minimum_value__316 ( day, region )
         valid_range :  <ARRAY>
         _FillValue :   3.402823e+38
         long_name :    Daily, Total-sky Averages
         units :        Watts per square meter
         hdf_name :     Longwave flux minimum value
         hdf_group :    Daily Averages/Total-Sky
         hdf_group_id : 316

The hdf_group and hdf_group_id attributes are only created if the file has duplicate variables
and groups are found. If there are duplicate variable but no groups (I don't know if such files
exist), sequential numbers are used instead of the group ids.
A problem which has not yet been resolved is that occasionally an SDS variable belongs to
more than one Vgroup. Right now NCL just lists such an SDS as belonging to the
first Vgroup of which it is a member.

3) Added caching for all HDF attributes, similar to the caching implemented for
the NetCDF module. This improves the performance of printing the contents of large
HDF files by several orders of magnitude.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@7854 10fe2890-8a40-4696-8b1d-7ccbb19c398e