This is to add the capability of specifying symbols as part of a dash pattern (so that, for example, one can draw dotted lines, lines formed from filled squares, lines made up of dots and dashes, and so on).
This was done in response to a user question, but it's something that has
been requested many times over the years.

Two routines have been added to the user interface: DPDPWS and DPDSYM.

The routine DPDPWS allows one to generate a character dash pattern in
which some of the characters are marked as symbol generators by having
their upper bits set.  Actually, it's pretty easy to do that with FORTRAN,
but some users won't know how, so I'm supplying a routine to make it easy

The routine CPDSYM draws the symbols: this version is capable of drawing
circles, squares, triangles, diamonds, and five-pointed stars, any of
which can be either solid or hollow.  This routine can easily be replaced
by a user to draw other symbols of his or her design.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@7925 10fe2890-8a40-4696-8b1d-7ccbb19c398e