This version of popRemap was created because SCD's firewall was causing problems for people outside the firewall.
The user *must* set an environment variable (NCL_POP_REMAP)
to tell popRemap where the desired remap file is located.

eg:   setenv NCL_POP_REMAP /fis/cgd/cseg/csm/mapping/maps

If the variables to be remapped are vector components,
a second environment variable (NCL_POP_REMAPV) must be
set. This specifies the location of the files containing
the rotation angles associated with the target grid.

eg:   setenv NCL_POP_REMAPV /cgd/cas/shea/POP_REMAP

If the these environment directories are not set,
the script will look in the current directory ("./")
for the files.

The above example directories are located behind the
firewall and are accesible by any machine behind the firewall.

Users outside the firewall must obtain the files and
place them in the same directory as the NCL script or
use the environment variable(s) where the files are
located.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@7945 10fe2890-8a40-4696-8b1d-7ccbb19c398e