BuiltInFuncs: modified all the conversion functions to behave better with respect to  _FillValues: If there is no fill value in the from variable and the conversion does not result in any out-of-range values for the to type, then no fill value is propagated.
If the function return value is temporary or the LHS has no fill value, then
the conversion func attempts to convert the fill value of the from variable.
If the conversion results in a valid value for the to type, it becomes the
fill value. If the conversion results in a out-of-range value for the to type,
or otherwise fails, then the default fill value for the to type is used.

If the LHS has a defined _FillValue attribute, then it is not changed.
Out-of-range or failed conversions are set to the value of the LHS _FillValue.

DataSupport: the stringtochar function now zero's out all memory in order
to prevent undefined values beyond the string length of the shorter strings.

NclTypebyte, NclTypechar, and NclTypeString:
The print format for the byte type is changed to 2 digit hexadecimal with '0x'
as a prefix for 4 characters total.

The print format for non-print and non-space characters is also changed to
the 4 digit hexadecimal format. However, character variables that are inserted
into strings are not affected by this.

NclVar.c: Fixed a problem with scalar variables being treated inconsistently.
The _FillValue attribute was added to array variables appropriately, but
in analogous situations it was not being added to scalar variables.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8037 10fe2890-8a40-4696-8b1d-7ccbb19c398e