This extends support for propagation of missing values tothe stringtochar and chartostring conversions. Because of thedimensional difference that occurs across the conversion there are some pecularities.
When chars are converted to string, the char missing value acts like
the default missing value, '\0', would. If there are 'good' characters
preceding the missing value character, they are converted to the
string; the missing value is changed to a '\0', and so the string
consists only of the characters preceding the missing value. If the
first element of one of the last dimensions of the character array is
the missing value then the behavior is different: The string value becomes
the character missing value converted to string. For example, if the
character missing value is 'd' and a 1D character array with characters
'd','e','f' is converted to string the string value will become
"d" with _FillValue set to "d". If it were assigned to a previously
defined scalar string variable with _FillValue set to "missing", it
would become "missing".

Going from string to character, elements of multi-D character arrays
beyond the len of the string are initially filled with the default
character missing value -- 0. This could be changed if the result
of the conversion is assigned to a pre-existing character variable
with a non-default missing value.
But note that a scalar string cannot be assigned to a pre-existing
1D character variable whose dimsize is other than the length of the
string + 1.
If a string's value is equal to its _FillValue, then all elements
of the corresponding character array are set to the character missing
value, but the length of the missing value string does not contribute
to finding the maximum length string that determines the size of the
character array upon conversion.

This is too much... I hope this subject goes away soon.
I have a script that illustrates these points.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8044 10fe2890-8a40-4696-8b1d-7ccbb19c398e