Updates to the asciiread function: Separated the numeric read function into 2 sections, one for float types (float, double), the other for integer types.
The float scan function now recognizes 'nan' (any case)
as a 'number'. Also because of system differences in the
implementation of the 'strtod' function, hexadecimal numbers
signified by an initial '0x' are explicitly NOT recognized for
float types. The initial '0' will be intepreted as a separate number.

The integer reading function does now recognize an initial '0x'
as identifying the hexadecimal representation of an integer.
Number can now be read correctly into 'byte' type variables.
For the byte type only, the octal representation of a number is
also understood: i.e., '077' would be intepreted as an octal number
when read as a byte type, but as a decimal number when read as any
of the other integer types. For the integer types digits following
a decimal point or an 'e' or'E' are ignored. A number starting with
a decimal point is given the value 0.

If a long value is out of range, it is silently replaced with the
maximum or minimum long value (+/-2147483647 for 32 bit systems).
If a double value is out of range, it is replaced by + or - inf.
Otherwise out-of-ranges values result in bit-wise truncation.
No error messages are produced.
(This has always been the case, and maybe should be revisited)


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8055 10fe2890-8a40-4696-8b1d-7ccbb19c398e