This is another change to fix the problem of contour fill bleeding. It modifies the MapTransformation so that it no longer erroneously leads EZMAP into thinking that points are outside the projection space when they are in fact only outside the data space.
It also modifies the contour rendering code such that if FixFillBleed is
set True it draw the line around the data space boundary in the opposite
direction. (As it turns out, this does not help much, but it is more
consistent with the way Conpack puts the line around the data space.)

This does fix TT# 047754.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8062 10fe2890-8a40-4696-8b1d-7ccbb19c398e