I am splitting up the "nfp" library into two pieces: the Fortran routines and the C wrapper routines.
This is due to the fact that PyNGL now depends on Fortran routines in
"nfp", and I don't want the build of PyNGL to be dependent on NCL
(which the C wrappers in the nfp library currently require).  By splitting
out the Fortran routines, they can be built standalone, and don't
need any NCL include files.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8098 10fe2890-8a40-4696-8b1d-7ccbb19c398e