Updated the yMakefile and MakeNcl.csh to reflect that the "nfpfort" library is now needed to build NCL.
I believe this is the last check-in with respect to moving the
Fortran files over.

Note to all: if you use the yMakefile scheme to update and build NCAR
Graphics and NCL, you will need to update your config directory along
with the directories under ni/src/lib and ni/src/ncl.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8101 10fe2890-8a40-4696-8b1d-7ccbb19c398e