Instead of the macros F_NO_UNDERSCORES and F_UPPERCASE, I have instead defined a new Template macro "NgCallF" that should be set to the value you want the NGCALLF macro to have. This macro is for defining whether a Fortran subroutine, when called from C, has an underscore appended, or is converted to uppercase.
The default value for NgCallF is "reg\#\#_", which assumes that
an underscore is appended after Fortran routine names.

If you want no munging of Fortran routine names, set
NgCallF to "reg" (without the double quotes) in the
appropriate config file.

If the Fortran routine name is converted to uppercase,
like with the Absoft ProFortran compiler, use "caps".

This change will effect the "c.h" file in
$NCARG/common/src/libncarg_c, which I will check in shortly.  It will
now be generated from a "c.h.sed" file, in which the SED_NGCALLF
string will be replaced with whatever "NgCallF" has been defined to.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8192 10fe2890-8a40-4696-8b1d-7ccbb19c398e