The "c.h" file is now generated from "c.h.sed". This is so the NGCALLF macro (which determines how Fortran subroutines need to be referenced when calling them from a C subroutine) can be set to the exact value that is appropriate for the compiler that it is being compiled with. Before, the NGCALLF macro was being set depending on just the system, which was not a good test.
For example, if you use the GNU compiler on a Linux system, there will
be underscores appended to Fortran routine names by default.  If you
use the "xlf" compiler on a Linux system, however, no underscores will
be appended.

A similar change will be made to the "ngmath.h" file shortly.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8193 10fe2890-8a40-4696-8b1d-7ccbb19c398e