For CLO: -x (echo) now works in both interactive and batch modes.
io.c changed to reflect that it no longer looks to see if NCL is running in
interactive mode to decide whether or not to echo commands.

ncl.l changed so its lexical analyzer outputs the currently-being-parsed text
if echo is desired.  Two conditions are met to output text:
    1. anything followed by a "\n"
    2. a "\n"

Meeting these conditions allows for conditional statements to be out prior to
their evaluation, which should aid in script debugging.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8212 10fe2890-8a40-4696-8b1d-7ccbb19c398e