This updates the NCL directory to build a new library called libnio.a, which contains NCL's IO layer without the parser, state machine, and built-in functions. It adds a few files, somewhat modified, from the HLU, LLU and ncarg_c libraries. This allows it to be linked without including any of these libraries or the GKS library. It does still depend on a few HLU and lower level include files. This will need to be sorted out if it is ever to be built independently of NCARG.
The HDF and NetCDF modules have been updated to allow
attribute strings to written over with a longer string when a file is
opened for writing. Formerly you were limited to a string no longer than
the current length. With large files, you may pay a performance
penalty for making a string longer, but no more than if you add a new
attribute to a file, something that has always been possible.

Infrastructure has been added to the File object in preparation for implementing
file options. This should have no effect on execution currently.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8219 10fe2890-8a40-4696-8b1d-7ccbb19c398e