Turned off optimization for gzclpo.f. This is in response to a problem that Dave Kennison had with creating a program that generates scrolling Star Wars titles. This program would seg fault if "gzclpo.f" was compiled with optimization ("-O") turned on.
I cut the last CVS message off too early (the one for the yMakefile in
the libncarg/polypack directory), but what I meant to say was that the
program seg faults if "ppinpo.f" and "gzclpo.f" are compiled with
optimization turned on. Dave said that since most of the other
routines in the polypack directory are similar to "ppinpo.f", they are
likely to have optimization problems as well.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8233 10fe2890-8a40-4696-8b1d-7ccbb19c398e