Adding a wrapper for the GENDAT routine, which generates a 2D array of "nice" random data. This routine is used in many low level NCAR Graphics examples, including many of the cpex0* examples.
Dennis, I have made the interface as follows:

 data = gendat((/ny,nx/), mlow:int, mhigh:int, dlow:numeric, dhigh:numeric)

where data[ny,nx] will be the size of the returned 2D array. "data"
will be double if dlow or dhigh are double, and float otherwise.  mlow
and mhigh indicate the approximate number of lows and highs in the
data (these must be between 1 and 25), and dlow and dhigh indicate the
exact minimum and maximum of the data.

Is this interface acceptable, or do you think it should be further
simplified to remove mlow and mhigh?  I think you at least
need to have the dimension sizes and dlow and dhigh.

Here's how it would be called for the famous "cpex08" example which
shows contours over Africa:

   z = gendat((/40,40/),15,15,-10.,110.)

Let me know if you envisioned this function differently.

--Mary


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8267 10fe2890-8a40-4696-8b1d-7ccbb19c398e