This check-in implements the procedure "setfileoption". It required a number of changes to the File class and the FileInterface class, including new stuctures to hold the options and their values, and new methods: set_option and initialize_file_record. The initialize method was required to allow options to be set before opening the file, as required by some of the options.
setfileoption has 3 arguments:

The first specifies either the file format as a string (e.g. "nc" for netcdf or
"grb" for GRIB -- actually any string that is a valid suffix), or else
it is a file object (the return value of addfile). If a format is specified
then the option only applies when a file of that type is first opened.

The second argument is a case-insensitive string naming the option.

The third argument is the option value, which can be of different types,
as required by the option. It can be a scalar or theoretically a 1D array,
although all currently implemented options take scalar values.

I have so far implemented 6 options, one for GRIB files, and 5 for NetCDF files.

The GRIB option name is "ThinnedGridInterpolation". It has the values
"linear" or "cubic", with "linear" as the default. It only
applies to certain GRIB files that have thinned
grids (called quasi-regular grids in the GRIB documentation). It can be
set on an open GRIB file, causing different values to be returned for the same
variable. Example (assuming the option has not been set):

v1 = f->GRIB_THINNED_GRID_VAR
setfileoption(f,"thinnedgridinterpolation","cubic")
v2 = f->GRIB_THINNED_GRID_VAR

v1 and v2 will almost certainly have many elements that are different, although
equivalent plots of each look very similar.

NetCDF options:

"Prefill" : True (default) or False. Valid only for files opened for writing.
If False, variables are not pre-filled with the fill values when defined.

"DefineMode" : False (default) or True. When True, NCL keeps the NetCDF library
in define mode without closing the file until variable data is read or written, or
until the option is set False again. This can have a very significant effect on
performance for files with lots of variables and/or attributes. The test case
that Dennis gave me goes from a running time of 8 minutes to around 10 seconds.
Essentially this provides the same kind of speedup as the "createfile" function,
without the necessity of learning a whole different (cumbersome) interface for
writing NetCDF files.

"HeaderReserveSpace" : positive integer (default is 0). This allows the user to reserve empty space
in the header area of the NetCDF file so that new attributes, dimensions, and variables
can be defined later without the significant performance penalty that usually occurs.

"SuppressClose" : False (default) or True. If True an open NetCdf file is not closed until
the user calls 'destroy' on the file object, NCL quits, or the option is set False again.
This may result in some additional speedup. It is really more of a research tool currently.
I would like to see if it improves DODS performance.

"Format" : "classic" (default), "64bitoffset" or "largefile". The last two are equivalent. This
option can only be set before a file is created, so you have to use the string format form
of the first argument. This is really format within a format. "classic" means standard 32bit
offset NetCDF. "64bitoffset" or "largefile" means create a file where the variables can be
larger than 2GIG (although this is not fully supported in NCL internally yet). When NetCDF4 is
implemented then it will be another option "netcdf4". This one probably should not be
documented until we actually can support >2GB variables.


CVS ----------------------------------------------------------------------


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8273 10fe2890-8a40-4696-8b1d-7ccbb19c398e