Modified the way the application object is created.  A default app object is only created once, and then if gsn_open_wks is called again, just a regular app is created.
This involved changing the parent of several of the data and
primitive objects from "noparent" or "defaultapp" to "appid", where
"appid" is retrieved via an Nhl function call.

This change is in response to a problem a user was having trying to
create multiple workstations.  The first workstation was okay, but when
he created subsequent workstations, it ended up clobbering the data object
of the first workstation, and hence the data wouldn't draw.

This doesn't solve another problem, however, and that is that the
app object itself gets clobbered when you call gsn_open_wks multiple
times.  I've sent email to Dave Brown about this, but meanwhile, I'm going
to make a change to gsn_open_wks (or else create a new routine) that
will allow you to create multiple workstations.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8278 10fe2890-8a40-4696-8b1d-7ccbb19c398e