Dave K looked into a problem that Christine Shields was having with the paleo_outline routine, and determined there were some things wrong with the Fortran driver that calls his SVBLED routine.
For one thing, the driver was assuming that the value "0" was a safe
value to use to "blank" out the double-size work array that SVBLED
depends on. This is not a good idea, because "0" could very well be
the mask value used to determine what is land and what is not (as we
found with a case that Christine gave us).

So, instead of setting this array initially to 0, it is now set
to 1-MSKVAL.

Secondly, Dave discovered that some incorrect values were being
used for "irgt" and "lonn", and gave instructions on how they
should be corrected. Instead of:

  irgt = 2 * nlon + 2
  lonn = 360

They are now set to:

  irgt = 2 * nlon
  lonn = real(lon(nlon))

This is consistent, also, with how it is done for the "lat" values.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8312 10fe2890-8a40-4696-8b1d-7ccbb19c398e