This implements raster fill mode for the LabelBar. LabelBar has a new resource called lbRasterFillOn. If set True the LabelBar boxes are drawn using a cell array rather than using polygon fill. Pattern fill is disabled in this case. Raster fill is not currently available if lbBoxSizing is set to "ExplicitSizing".
ContourPlot forces the LabelBar to use raster fill if the cnFillMode
is set to RasterFill (or cnRasterModeOn is True).

There will be no apparent difference to the LabelBar rendering except
on certain PS printers that have slight color variations for the same
color drawn using a fill area vs an image.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8345 10fe2890-8a40-4696-8b1d-7ccbb19c398e