Dave B fixed the infamous problem with the labelbar colors not matching the contour colors when a raster contour fill was being done and the output being sent to a PS printer. (This didn't occur for all printers, and thus we consider this a printer bug, not an NCL bug.)
The fix was to force the labelbar to also be done in raster mode (i.e.
as an image rather than a polygon with a fill color) if the contours
are in raster mode.

If you create your own labelbar, however, you will have to set
lbRasterFillOn to True if contour fill is raster. That is the change
being made here.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8354 10fe2890-8a40-4696-8b1d-7ccbb19c398e