1) the byte swapping option was not working with sequential fortran files because the control word was not getting byte-swapped along with the data. I fixed this, but it needs testing.
2) Also, I added an error message for the situation where the user
is trying to write to a particular record that already exists in the
Fortran file. If the record size is not the same size as the
variable being written, the code was just returning a fatal error
without explanation. I have fixed this and now the following message is
returned:
fbinrecwrite: data variable size does not match record size

3)And more seriously, I found a bug in writing to a previously
existing record (fbinrecwrite("test.bin",3,var))
The file pointer was not being set correctly in this case, and
the file was being trashed. I have fixed this.
Hard to believe this has never been encountered, but I guess people
usually just append using -1 as the second argument.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8391 10fe2890-8a40-4696-8b1d-7ccbb19c398e