Fixes for problem of NCL dumping core if an invalid filename is provided on the command line.
Ncl.c: no longer assumes a ".ncl" extention to a file; file argument properties
are more closely examined to see if the argument is indeed a file.  If not, it's
treated as a variable.

ncl.l: now handles the case whereby a legitimate filename is provided to NCL, but
permissions are denied (for whatever reason) for the user to access the file, and
a null filename isn't strcpy()'d.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8435 10fe2890-8a40-4696-8b1d-7ccbb19c398e