This is a new function for performing canonical correlation analysis between two sets of variables.
This wrapper was written to handle multiple dimensions, but then Dennis
decided it should only handle the minimum (2 dimensions). However, I'm
going to leave it as is, because it doesn't really really hurt anything.
The wrapper function where the arguments are actually registered have been
changed to expect only 2 dimensions.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8472 10fe2890-8a40-4696-8b1d-7ccbb19c398e